+++ file.h
@@ -1,4 +1,4 @@
 #pragma once
 
-bool file_load(const char* filename, void** data, uint64_t* data_size);
-bool file_load_str(const char* filename, char** data, uint64_t* str_len);
+bool file_load(const char* filename, void** data, size_t* data_size);
+bool file_load_str(const char* filename, char** data, size_t* str_len);
